refactor(v2): isolate legacy defaults and split workspace surfaces
