+++ apps/server/src/routes/sources/$mediaSourceId/components/source-media-page.tsx
@@ -1,16 +1,22 @@
 import { Button } from "@solid-imager/ui/button";
+import type { SearchPersistenceSurface } from "@solid-imager/ui/hooks/use-current-search-persistence";
 import { createPresetClient } from "@solid-imager/ui/preset-client";
 import { sourceMediaQueryKeys } from "@solid-imager/ui/query-options";
 import { RouteDataPendingScreen } from "@solid-imager/ui/router-status";
+import type { SourceMediaScreenProps } from "@solid-imager/ui/screens/source-media-screen.types";
 import { SourceMediaPage as SourceMediaPageComponent } from "@solid-imager/ui/source-media-page";
 import { createQuery, useQueryClient } from "@tanstack/solid-query";
-import { useLocation, useNavigate, useParams } from "@tanstack/solid-router";
-import { type Accessor, createSignal, onMount, Show } from "solid-js";
+import { useParams } from "@tanstack/solid-router";
+import {
+	type Accessor,
+	type Component,
+	createSignal,
+	onMount,
+	Show,
+} from "solid-js";
 import { BulkActionDialog } from "~/components/media/bulk-action-dialog";
 import { MediaGridItem } from "~/components/media/media-grid-item";
 import { MoveCopyMediaDialog } from "~/components/media/move-copy-media-dialog";
-import { ThumbnailImage } from "~/components/media/thumbnail-image";
-import { UploadMediaModal } from "~/components/upload-media-modal";
 import { createServerTransport } from "~/hooks/use-media-source-events";
 import { PresetClient as rawPresetClient } from "~/infrastructure/api/clients/preset-client";
 import { startDownloadJobs } from "~/infrastructure/api-clients/downloads-api";
@@ -44,12 +50,20 @@ import {
 
 const PresetClient = createPresetClient(rawPresetClient);
 
-export function SourceMediaPage(
-	props: { mediaSourceId?: Accessor<string>; variant?: "default" | "v2" } = {},
+export type SourceMediaPageControllerProps = {
+	mediaSourceId?: Accessor<string>;
+	routeVersion: "default" | "v2";
+	screenComponent: Component<SourceMediaScreenProps>;
+	uploadModalComponent: SourceMediaScreenProps["uploadModalComponent"];
+	persistenceSurface: SearchPersistenceSurface;
+	onOpenMediaDetail?: SourceMediaScreenProps["onOpenMediaDetail"];
+	renderMediaPreview?: SourceMediaScreenProps["renderMediaPreview"];
+};
+
+export function SourceMediaPageController(
+	props: SourceMediaPageControllerProps,
 ) {
 	const params = useParams({ strict: false });
-	const location = useLocation();
-	const navigate = useNavigate();
 	const mediaSourceId = () => props.mediaSourceId?.() ?? params().mediaSourceId;
 	const queryClient = useQueryClient();
 	const [isMounted, setIsMounted] = createSignal(false);
@@ -102,13 +116,13 @@ export function SourceMediaPage(
 					title="メディア一覧"
 				/>
 			}
-			when={props.variant === "v2" || isMounted()}
+			when={props.routeVersion === "v2" || isMounted()}
 		>
 			<SourceMediaPageComponent
 				enableVirtualization
-				variant={props.variant}
 				mediaSourceId={mediaSourceId}
 				mediaSourceName={mediaSourceName}
+				persistenceSurface={props.persistenceSurface}
 				transport={transport}
 				presetClient={PresetClient}
 				actions={{
@@ -141,12 +155,18 @@ export function SourceMediaPage(
 				onClearSelection={handleCancelSelect}
 				selectedCount={() => selectedMediaIds().length}
 				onEnterBulkSelectMode={() => setIsBulkSelectMode(true)}
+				screenComponent={props.screenComponent}
+				scrollContainerSelector={
+					props.routeVersion === "v2"
+						? `[data-media-scroll="${mediaSourceId() ?? "v2-source"}"]`
+						: undefined
+				}
 				renderItem={(media, options) => (
 					<MediaGridItem
 						imageLoadPolicy={options.imageLoadPolicy}
 						media={media}
 						priority={options.priority}
-						routeVersion={props.variant === "v2" ? "v2" : "default"}
+						routeVersion={props.routeVersion}
 						onContextMenu={options.onContextMenu}
 						isBulkSelectMode={options.isBulkSelectMode}
 						isSelected={options.isSelected || options.isPreviewSelected}
@@ -154,36 +174,21 @@ export function SourceMediaPage(
 						onToggleSelect={() => handleToggleSelect(media.id)}
 					/>
 				)}
-				onOpenMediaDetail={(media) => {
-					sessionStorage.setItem("v2:media-return", location().href);
-					void navigate({
-						params: {
-							mediaId: media.id,
-							mediaSourceId: media.mediaSourceId,
-						},
-						to: "/v2/sources/$mediaSourceId/$mediaId",
-					});
-				}}
-				renderMediaPreview={(media) => (
-					<ThumbnailImage
-						alt={media.fileName}
-						class="h-full w-full object-contain"
-						height={media.height}
-						loading="eager"
-						media={media}
-						requestedSize={512}
-						width={media.width}
-					/>
-				)}
+				onOpenMediaDetail={props.onOpenMediaDetail}
+				renderMediaPreview={props.renderMediaPreview}
 				moveCopyDialogComponent={MoveCopyMediaDialog}
-				uploadModalComponent={UploadMediaModal}
+				uploadModalComponent={props.uploadModalComponent}
 				showOpenInNewTab
 			/>
 
 			{/* 一括選択ツールバー */}
 			<Show when={isBulkSelectMode()}>
 				<div
-					class="fixed bottom-[calc(1rem+env(safe-area-inset-bottom))] left-1/2 z-50 flex w-[calc(100%-2rem)] max-w-md -translate-x-1/2 flex-wrap items-center justify-center gap-2 rounded-2xl border border-primary/20 bg-background/95 px-3 py-3 shadow-lg backdrop-blur sm:bottom-[calc(1.5rem+env(safe-area-inset-bottom))] sm:w-auto sm:max-w-none sm:flex-nowrap sm:gap-4 sm:rounded-full sm:px-6"
+					class={
+						props.routeVersion === "v2"
+							? "fixed bottom-[calc(1rem+env(safe-area-inset-bottom))] left-1/2 z-50 flex w-[calc(100%-2rem)] max-w-md -translate-x-1/2 flex-wrap items-center justify-center gap-2 rounded-md border border-[var(--v2-border)] bg-[var(--v2-surface)] px-3 py-3 sm:bottom-[calc(1.5rem+env(safe-area-inset-bottom))] sm:w-auto sm:max-w-none sm:flex-nowrap sm:gap-3 sm:px-4"
+							: "fixed bottom-[calc(1rem+env(safe-area-inset-bottom))] left-1/2 z-50 flex w-[calc(100%-2rem)] max-w-md -translate-x-1/2 flex-wrap items-center justify-center gap-2 rounded-2xl border border-primary/20 bg-background/95 px-3 py-3 shadow-lg backdrop-blur sm:bottom-[calc(1.5rem+env(safe-area-inset-bottom))] sm:w-auto sm:max-w-none sm:flex-nowrap sm:gap-4 sm:rounded-full sm:px-6"
+					}
 					data-testid="bulk-actions-bar"
 				>
 					<span class="w-full text-center font-medium text-sm sm:w-auto">
