+++ apps/server/src/routes/v2/manager.tsx
@@ -1,6 +1,8 @@
 import { useManagerPage } from "@solid-imager/ui/hooks/use-manager-page";
-import { ManagerScreen } from "@solid-imager/ui/screens/manager-screen";
-import type { V2ManagerTransferFormat } from "@solid-imager/ui/screens/v2-manager-screen";
+import {
+	V2ManagerScreen,
+	type V2ManagerTransferFormat,
+} from "@solid-imager/ui/screens/v2-manager-screen";
 import { toast } from "@solid-imager/ui/toast";
 import { useQueryClient } from "@tanstack/solid-query";
 import { createFileRoute } from "@tanstack/solid-router";
@@ -161,10 +163,6 @@ function V2ManagerRoute() {
 	});
 
 	return (
-		<ManagerScreen
-			manager={manager}
-			transferActions={transferActions}
-			variant="v2"
-		/>
+		<V2ManagerScreen manager={manager} transferActions={transferActions} />
 	);
 }
