+++ apps/server/src/routes/v2/search.tsx
@@ -1,7 +1,10 @@
-import { useCurrentSearchPersistence } from "@solid-imager/ui/hooks/use-current-search-persistence";
+import {
+	persistSearchScrollPosition,
+	useCurrentSearchPersistence,
+} from "@solid-imager/ui/hooks/use-current-search-persistence";
 import { useSearchPage } from "@solid-imager/ui/hooks/use-search-page";
 import { createPresetClient } from "@solid-imager/ui/preset-client";
-import { SearchScreen } from "@solid-imager/ui/screens/search-screen";
+import { V2SearchScreen } from "@solid-imager/ui/screens/v2-search-screen";
 import {
 	createFileRoute,
 	useLocation,
@@ -42,7 +45,9 @@ export const Route = createFileRoute("/v2/search")({
 function V2SearchRoute() {
 	const location = useLocation();
 	const navigate = useNavigate();
-	const isSearchStateRestored = useCurrentSearchPersistence("all");
+	const isSearchStateRestored = useCurrentSearchPersistence("all", {
+		surface: "v2",
+	});
 	const page = useSearchPage({
 		searchMedia,
 		searchSimilar,
@@ -62,7 +67,10 @@ function V2SearchRoute() {
 		// so scrolling does not stop for another request every two or three rows.
 		limit: () => Math.max(searchState.limit, V2_SEARCH_RESULTS_PER_PAGE),
 		scrollY: () => searchState.scrollY,
-		setScrollY: (value) => setSearchState("scrollY", value),
+		setScrollY: (value) => {
+			setSearchState("scrollY", value);
+			persistSearchScrollPosition("all", value, { surface: "v2" });
+		},
 		setOffset: (value) => setSearchState("offset", value),
 		mode: () => searchState.mode,
 		similarityAnchorMediaId: () => searchState.similarityAnchorMediaId,
@@ -82,7 +90,7 @@ function V2SearchRoute() {
 	});
 
 	return (
-		<SearchScreen
+		<V2SearchScreen
 			enableVirtualization
 			filterData={page.filterData}
 			onSelectSource={(id) => setSearchState("selectedSource", id)}
@@ -122,7 +130,6 @@ function V2SearchRoute() {
 			selectedSource={searchState.selectedSource}
 			ssrGuard
 			sources={page.sources()}
-			variant="v2"
 		/>
 	);
 }
