+++ apps/tauri/src/routes/search.tsx
@@ -1,4 +1,5 @@
 import { Button } from "@solid-imager/ui/button";
+import { persistSearchScrollPosition } from "@solid-imager/ui/hooks/use-current-search-persistence";
 import { useSearchPage } from "@solid-imager/ui/hooks/use-search-page";
 import { createPresetClient } from "@solid-imager/ui/preset-client";
 import { SearchScreen } from "@solid-imager/ui/screens/search-screen";
@@ -61,7 +62,10 @@ function SearchRoute() {
 		sortOrder: () => searchState.sortOrder,
 		limit: () => searchState.limit,
 		scrollY: () => searchState.scrollY,
-		setScrollY: (y) => setSearchState("scrollY", y),
+		setScrollY: (y) => {
+			setSearchState("scrollY", y);
+			persistSearchScrollPosition("all", y);
+		},
 		setOffset: (o) => setSearchState("offset", o),
 		mode: () => searchState.mode,
 		similarityAnchorMediaId: () => searchState.similarityAnchorMediaId,
