+++ packages/ui/src/hooks/use-current-search-persistence.test.ts
@@ -6,6 +6,8 @@ import {
 	setSearchState,
 } from "../stores/search-store";
 import {
+	persistSearchScrollPosition,
+	type SearchPersistenceOptions,
 	type SearchPersistenceSource,
 	useCurrentSearchPersistence,
 } from "./use-current-search-persistence";
@@ -91,9 +93,10 @@ describe("useCurrentSearchPersistence", () => {
 
 	const mountPersistence = (
 		sourceId: SearchPersistenceSource,
+		options?: SearchPersistenceOptions,
 	): MountedPersistence => {
 		const mounted = createRoot((dispose) => {
-			const isRestored = useCurrentSearchPersistence(sourceId);
+			const isRestored = useCurrentSearchPersistence(sourceId, options);
 			return { dispose, initialValue: isRestored(), isRestored };
 		});
 		mountedRoots.push(mounted);
@@ -193,7 +196,7 @@ describe("useCurrentSearchPersistence", () => {
 	});
 
 	it("preserves the current scroll position while restoring search state", async () => {
-		setSearchState("scrollY", 1840);
+		sessionStorage.setItem("search-scroll:legacy:current-all", "1840");
 		sessionStorage.setItem(
 			"current-all",
 			JSON.stringify(createPersistedSimpleState("saved query")),
@@ -203,6 +206,38 @@ describe("useCurrentSearchPersistence", () => {
 		await flushMicrotasks();
 
 		expect(searchState.scrollY).toBe(1840);
+	});
+
+	it("keeps v2 scroll persistence separate from the legacy surface", async () => {
+		vi.useFakeTimers();
+		sessionStorage.setItem(
+			"v2:current-all",
+			JSON.stringify(createPersistedSimpleState("shared query")),
+		);
+		sessionStorage.setItem(
+			"current-all",
+			JSON.stringify(createPersistedSimpleState("legacy query")),
+		);
+		sessionStorage.setItem("search-scroll:legacy:current-all", "240");
+		sessionStorage.setItem("search-scroll:v2:current-all", "1840");
+
+		mountPersistence("all", { surface: "v2" });
+		await flushMicrotasks();
+
+		expect(searchState.searchQuery).toBe("shared query");
+		expect(searchState.scrollY).toBe(1840);
+
+		setSearchState("scrollY", 2200);
+		expect(searchState.scrollY).toBe(2200);
+		persistSearchScrollPosition("all", searchState.scrollY, { surface: "v2" });
+		await flushMicrotasks();
+		await vi.advanceTimersByTimeAsync(1000);
+
+		expect(sessionStorage.getItem("search-scroll:v2:current-all")).toBe("2200");
+		expect(sessionStorage.getItem("search-scroll:legacy:current-all")).toBe(
+			"240",
+		);
+		expect(sessionStorage.getItem("current-all")).toContain("legacy query");
 	});
 
 	it("does not inherit a source for legacy vector sessions", async () => {
