+++ packages/ui/src/sort-controls.tsx
@@ -1,3 +1,4 @@
+import { Label } from "./label";
 import {
 	Select,
 	SelectContent,
@@ -39,7 +40,7 @@ export function SortControls(props: SortControlsProps) {
 	return (
 		<div class={props.className}>
 			<div class="space-y-2">
-				<p class="font-medium text-sm leading-none">ソート</p>
+				<Label>ソート</Label>
 				<div class="grid grid-cols-2 gap-2">
 					<Select
 						itemComponent={(itemProps) => (
@@ -56,7 +57,7 @@ export function SortControls(props: SortControlsProps) {
 						placeholder="項目"
 						value={props.sortBy}
 					>
-						<SelectTrigger aria-label="ソート項目">
+						<SelectTrigger>
 							<SelectValue<string>>
 								{(state) =>
 									getSortLabel(
@@ -82,7 +83,7 @@ export function SortControls(props: SortControlsProps) {
 						placeholder="順序"
 						value={props.sortOrder}
 					>
-						<SelectTrigger aria-label="ソート順">
+						<SelectTrigger>
 							<SelectValue<string>>
 								{(state) =>
 									state.selectedOption() === "asc" ? "昇順" : "降順"
